requirements.txt
alabaster==0.7.8
anaconda-client==1.4.0
anaconda-navigator==1.2.1
aniso8601==1.2.0
AnyQt==0.0.8
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
asn1crypto==0.23.0
astropy==1.2.1
aws==0.2.5
awscli==1.6.2
AWSIoTPythonSDK==1.2.0
Babel==2.3.3
backports-abc==0.4
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
bcdoc==0.12.2
bcrypt==3.1.4
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.0
boto==2.33.0
boto3==1.4.0
botocore==0.73.0
bottle==0.12.13
Bottleneck==1.1.0
cdecimal==2.3
certifi==2015.4.28
cffi==1.11.2
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
cocoex==0.0.0
colorama==0.2.5
conda==4.2.13
conda-build==1.21.3
configobj==5.0.6
configparser==3.5.0b2
contextlib2==0.5.3
cryptography==2.1.2
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
deap==1.0.2
decorator==4.0.9
dill==0.2.5
docopt==0.6.2
docutils==0.13.1
dynd==0.7.3.dev1
enum34==1.1.6
et-xmlfile==1.0.1
ete2==2.3.10
Fabric==1.14.0
fastcache==1.0.2
feedparser==5.2.1
filechunkio==1.8
Flask==0.11.1
Flask-Ask==0.8.2
Flask-Cors==2.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
genson==0.1.0
gevent==1.1.1
google-api-python-client==1.5.3
GraphLab-Create==2.1
GraphLab-Create-License==2.1
greenlet==0.4.10
grin==1.2.1
h5py==2.6.0
HeapDict==1.0.0
httplib2==0.9.2
idna==2.1
imagesize==0.7.1
ipaddress==1.0.16
ipykernel==4.3.1
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jmespath==0.5.0
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==2.1.0
mistune==0.7.2
mock==2.0.0
mongo==0.2.0
mongopy==0.1
mpmath==0.19
mrjob==0.5.6
multipledispatch==0.4.9
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.13.3
oauth2client==3.0.0
odo==0.5.0
openpyxl==2.3.2
pandas==0.18.1
paramiko==2.3.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==1.10.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
pipreqs==0.4.9
ply==3.8
prettytable==0.7.2
prompt-toolkit==1.0.3
protobuf==3.1.0
psclient==2.0
psutil==4.3.0
ptyprocess==0.5.1
pudb==2017.1.2
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.2.3
PyForms==0.1.7.3
Pygments==2.1.3
pymongo==3.3.0
PyNaCl==1.1.2
pyOpenSSL==16.0.0
pyparsing==2.2.0
pysettings==0.1
pysolr==3.6.0
pyswarm==0.6
pytest==2.9.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
QtPy==1.0.2
quora==0.1.23
redis==2.10.5
requests==2.13.0
rope==0.9.4
rsa==3.1.2
ruamel-yaml===-VERSION
s3transfer==0.1.4
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.1
scoop==0.7.1.1
selenium==3.0.2
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
spyder==2.3.9
SQLAlchemy==1.0.13
sseclient==0.0.8
statsmodels==0.6.1
subprocess32==3.2.7
sympy==1.0
tables==3.2.2
terminado==0.6
textblob==0.11.1
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
twilio==5.6.0
unicodecsv==0.14.1
Unidecode==0.4.19
uritemplate==0.6
urwid==1.3.1
watson-developer-cloud==0.25.1
wcwidth==0.1.7
Werkzeug==0.11.10
wordcloud==1.3.1
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
yarg==0.1.9
